Rename preview finders
Rename preview finder functions to abstract them from Sketch
implementation names. As discovered by @TomasLinhart in #8, Sketch 3
changed the names of some messages. By generalizing the naming used in
the plugin, we should be able to support Sketch 2 and 3 with less
confusion, and we're better insulated from future Sketch changes.