Support Sketch 3
This change adds support for Sketch 3, building on the work done in #8.
The main difference is to use `respondsToSelector:` to send the
appropriate message based on what's available in the executing version
of Sketch. We also more gracefully handle the case where layer
collection messages disappear, so the subsequent search strategy will
still have a chance to run.

Thanks to @TomasLinhart for his work.