Log debugging information
Add the ability to selectively log debug info. This should help
troubleshoot future problems with Artboard selection. It seems that in
the latest version of Sketch 2, a layer within an Artboard must be
selected for that Artboard to register as selected. I think this is a
new development, but I have no way to test it. Regardless Artboard
selection works as expected in Sketch 3.