Create preview size config plugin
Prompt the user for their desired preview size, and write their
selection out to a plist in the cache directory. Less transient
locations seem to not play well with sandboxing, so given the ease of
re-selecting the desired preview size, this is good enough.

It might also be desireable to query Sketch for the possible preview
sizes, but we'll do that if/when the possible values ever change.